+++ R/reexport-meta.R
@@ -1,7 +1,0 @@
-#' @importFrom meta pairwise
-#' @export
-meta::pairwise
-
-#' @importFrom meta backtransf
-#' @export
-meta::backtransf